src/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-PyMongo==2.3.0
pymongo==3.12.0
SQLAlchemy==1.4.22
dnspython==1.16.0
validators==0.18.2
mysql-connector-python==8.0.26
